Introduced Teff
Added Teff dependance. Fit power law, use the power determined as the starting point for the dependance in the mean function.
Currently, it makes worse predictions than when it is not included. May need to mess about with length sclaes and sigmas, i.e. add a third length scale and sigma.

Or it might be due to the factor our front, A. 

Or something completely different B)