Improved Mass Age GP. Added BV Mass Age GP
Altered GP - improved resid and Z plots for Mass,Age GP. better values for the hyper parameters (I think) and better prior guesses on power laws. Residual plot is very straight, but I think the std is still a little large.

Added a third GP that uses mass, age and B-V. It is looking promising, residual fits well for lower 85% of rotation range, above this it seems to over predicit. Z plot errors are slightly better than the  one from mass,age, but not perfect as its a little large (about 2x higher than it should be at the peak)