Added plot for fit
Ran the code for all data using current model and included the fit here.
The model seems to perform poorly at the upper and lower limits, this suggests that there may be an issue with the model. In the middle region, the model is under-predicting. 
The error is also getting massive in the mid -> upper limit of rotation/age. 
This makes me think that I am maybe adding error poorly, plotting a scatter of age vs prot with error bars I can see that the larger values have a far greater error than the lower ones