removed latent rot and Mass dependance fixed noise
The mean function was being dominated by BV such that the mass funciton had little to no effect on the fit. 
Not sure why I had a latent parameter for rotation, not sure if I should or not. It will be easy to add in if neccesary but for now it fits well. 
Fixed a bug wherein noise was only being added to the data never subtracted (i.e data + noise not data +/- noise)
Generally tidying up