Added metalicty, improved plots
I have added in a dependance on metalicity, suggested by Guy and after reading the following paper. Seems to account for a lot of the previous outlier points.
(Evidence for Metallicity-Dependant Spin Evolution in the Kepler field Louis Amard et al.)

My straight line predicted vs data plot was, up to this commit, wrong. I was ordering the data which pretty much just took all meaning out of the plot and made it always appear that the GP was performing well, when in fact it wasnt. Fixing this required a few tweaks to improve the overall fit, but has now been improved.
As such the Z plot also required a few small bug fixes.